update intellijs warnings

+++ user_provided/org/eclipse/titan/LaravelTestTTCN3/user_provided/PacketHeaderDescr.java
@@ -49,7 +49,7 @@ public class PacketHeaderDescr {
 		long m_length = 0;
 		for (int i = 0; i < nr_bytes_in_length; i++) {
 			long shift_count = byte_order == HeaderByteOrder.Header_MSB ? nr_bytes_in_length - 1 - i : i;
-			m_length |= buffer_pointer[(int) (length_offset + i)] << (8 * shift_count);
+			m_length |= (long) buffer_pointer[(int) (length_offset + i)] << (8 * shift_count);
 		}
 		m_length *= length_multiplier;
 		if (value_offset < 0 && (long)m_length < -value_offset) {
